dashboard: add replication status panel
tnt_replication_status was introduced in metrics 0.13.0. This patch adds
replication status timeseries panel to cluster overview section and
corresponding "stream" label pairs to Telegraf.

Timeseries is a base panel since Grafana 8.0. It is possible to set
value mapping in a timeseries panel, which is convenient for replication
status display. There is no way to make a 8.x timeseries panel with
grafana/grafonnet-lib tools for now (7.x subfolder graph
is incompatible), so this patch also introduces a rough draft of a
timeseries panel template.

Grafana requirement was bumped to 8.0.0.

Part of #133