dashboard: add panels for expirationd module statistics
Statistics for expirationd module was introduced in
expirationd 1.2.0 [1]. Statistics integrated with metrics and enabled
by default.

To disable statistics integrated with metrics, call:

expirationd.cfg({
    metrics = false
})

This patch adds panels for counters:
- expirationd_checked_count
- expirationd_expired_count
- expirationd_restarts
- expirationd_working_time

The meaning of counters is same as for expirationd.stats() [2].

expirationd panels are stored in "expirationd module statistics"
section. Mostly it copy-paste from TDG's dashboard [3][4].

1. https://github.com/tarantool/expirationd/releases/tag/1.2.0
2. https://tarantool.github.io/expirationd/#stats
3. https://github.com/tarantool/grafana-dashboard/commit/ca6e0e192fade0bb5abfd4cf1e4f2ffecb04e464
4. https://github.com/tarantool/grafana-dashboard/blob/0c623e0fae8e526976ed70da5e5f5a6640856275/dashboard/panels/tdg/expirationd.libsonnet

Closes #149