dashboard: separate datasource and datasource type
Before this patch, datasource request target choice was based on
a datasource value. Only two possible values were supported for
a datasource: ${DS_INFLUXDB} for a templating InfluxDB  datasource and
${DS_PROMETHEUS} for a templating Prometheus datasource. After this
patch, datasource type to choose a target is set with a separate
variable, so now it's possible to pass a non-templating datasource
value.

Part of #90