dashboard: use fill(null) in InfluxQL requests
This patch replaces fill(none) with fill(null) in dashboards InfluxQL
requests. After this patch, all panels with non-derivative requests
will show the gap in metrics for an interval if there were no metrics
for this interval. Before this patch, they had displayed a straight line
connecting border values in the area. Derivative requests will still
display a straight line [1]. There is a workaround to make derivative
requests result consistent with non-derivative one, but it makes query
more complicated both in terms of readability and performance, so it's
not worth it.

1. https://github.com/influxdata/influxdb/issues/7185

Closes #23