doc: move list of supported metrics to a file
Move list of supported metrics from wiki [1] to a file. It is
inconvenient to work with wiki since one should manually update the list
after merge. It is also easy to miss a presence of wiki page and this
approach is not portable. By using simple markdown file to store
supported metrics info, one may update list in the same commit as
new metrics support itself.

1. https://github.com/tarantool/grafana-dashboard/wiki/Metrics-support