dashboard: use Grafana 11 color scheme in overview table
Before this patch, older Grafana 6 color scheme were used for overview
panel. It was fine for Grafana from 8 to 10, but don't work properly
with Grafana 11. After this patch, the panel works properly with
Grafana 11 and newer. For Grafana 9 and 10 everything works fine, except
for row coloring: only single cell is colored [1]. For Grafana 8 and
older, table column hiding no longer works properly so it's simpler
for us to treat it as unsupported since it's EOLed 1 year and 3 months
ago [2].

1. https://github.com/grafana/grafana/issues/27542
2. https://endoflife.date/grafana

Follows #234